+++ src/lib/assets/i18n/ast.ts
@@ -4,6 +4,7 @@ export const locale = {
 	data: {
 		BACK: 'tornar',
 		CONTINUE: 'siguir',
-		SUBMIT: 'unviar'
+		SUBMIT: 'unviar',
+		IN_ORDER: 'correuto'
 	}
 };