+++ src/lib/assets/i18n/zh.ts
@@ -4,6 +4,7 @@ export const locale = {
 	data: {
 		BACK: '返回',
 		CONTINUE: '继续',
-		SUBMIT: '提交'
+		SUBMIT: '提交',
+		IN_ORDER: '已完成'
 	}
 };